requirements.txt
onnx
onnxruntime-gpu
onnxoptimizer
cv2-python
onnxconverter-common
onnxruntime-tools
numpy
tqdm